Fix bug with installation test
