+++ tests/torch/test_compression_training.py
@@ -22,9 +22,9 @@ from copy import deepcopy
 from pytest import approx
 
 from examples.torch.common.utils import get_name
-from tests.common.helpers import TEST_ROOT
+from tests.common.helpers_without_dependencies import TEST_ROOT
 from tests.torch.helpers import Command
-from tests.common.helpers import get_cli_dict_args
+from tests.common.helpers_without_dependencies import get_cli_dict_args
 from tests.torch.test_sanity_sample import create_command_line, update_compression_algo_dict_with_legr_save_load_params
 # pylint: disable=redefined-outer-name
 # sample
