+++ tests/torch/test_config_schema.py
@@ -6,7 +6,7 @@ import jsonschema
 import pytest
 
 from nncf.config import NNCFConfig
-from tests.common.helpers import PROJECT_ROOT, TEST_ROOT
+from tests.common.helpers_without_dependencies import PROJECT_ROOT, TEST_ROOT
 
 GOOD_CONFIG_SOURCES = [
     PROJECT_ROOT / Path("examples/torch/classification/configs"),
