+++ tests/torch/test_extensions_build.py
@@ -6,7 +6,7 @@ import shutil
 
 import torch
 
-from tests.common.helpers import TEST_ROOT
+from tests.common.helpers_without_dependencies import TEST_ROOT
 from tests.torch.helpers import Command
 
 EXTENSIONS_BUILD_FILENAME = 'extensions_build_checks.py'
