+++ tests/torch/test_install.py
@@ -12,7 +12,7 @@
 """
 
 import pytest
-from tests.common.helpers import run_install_checks
+from tests.common.helpers_without_dependencies import run_install_checks
 
 
 @pytest.fixture(name="venv_type",
